Moved types into repository
